Python_Codes/requirements.txt
opencv-python
# inference==0.9.17
supervision>=0.20.0
ultralytics
pyserial
cvzone
